requirements.txt
requests
toga-android>=0.4.0
